+++ civis/polling.py
@@ -120,18 +120,20 @@ class PollableResult(CivisAsyncResultBase):
             self._last_polled = time.time()
         self._last_result = None
 
-        self._polling_thread = _ResultPollingThread(self._check_result, (),
-                                                    polling_interval)
+        self._begin_tracking()
+
+    def _begin_tracking(self):
+        """Start monitoring the Civis Platform job"""
+        with self._condition:
+            if getattr(self, 'poller', None) is None:
+                raise RuntimeError('Internal error: Must set polling '
+                                   'function before initializing thread.')
+            self._reset_polling_thread(self.polling_interval)
 
     def _check_result(self):
         """Return the job result from Civis. Once the job completes, store the
         result and never poll again."""
         with self._condition:
-            # Start a single thread continuously polling.
-            # It will stop once the job completes.
-            if not self._polling_thread.is_alive() and self._result is None:
-                self._polling_thread.start()
-
             if self._result is not None:
                 # If the job is already completed, just return the stored
                 # result.
@@ -190,8 +192,10 @@ class PollableResult(CivisAsyncResultBase):
     def _reset_polling_thread(self,
                               polling_interval=_DEFAULT_POLLING_INTERVAL):
         with self._condition:
-            if self._polling_thread.is_alive():
+            if (getattr(self, '_polling_thread', None) is not None and
+                    self._polling_thread.is_alive()):
                 self._polling_thread.cancel()
             self.polling_interval = polling_interval
             self._polling_thread = _ResultPollingThread(self._check_result, (),
                                                         polling_interval)
+            self._polling_thread.start()