+++ CHANGELOG.md
@@ -29,6 +29,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   with a 404 status code. This fix affects the executors and joblib backend, which
   use the ``ContainerFuture``.
 - Tell ``flake8`` to ignore a broad except in a ``CivisFuture`` callback.
+- When unpickling ``ModelFuture`` objects, set the polling function before starting the polling thread.
+- Always subscribe to PubNub notifications (if using notifications) before
+  polling for status for the first time. This can avoid rare instances of
+  missing job completion notifications.
 
 ### Added
 - ``civis.resources.cache_api_spec`` function to make it easier to record the